+++ example/lib/test_case.dart
@@ -16,86 +16,74 @@ class AesTestCase extends TestCase {
   final String plain;
   final String expectedEncrypted;
 
-  AesTestCase(String name, this.aes, this.plain, this.expectedEncrypted) : super(name);
+  AesTestCase(String name, this.aes, this.plain, this.expectedEncrypted)
+      : super(name);
 
   @override
   Future<bool> execute() async {
-    Uint8List encrypted = await aes.encryptIsolated(hex.decode(plain));
-    Uint8List decrypted = await aes.decryptIsolated(encrypted);
-    return (hex.encode(encrypted) == expectedEncrypted) && (hex.encode(decrypted) == plain);
+    return validateEncrypt();
   }
 
   Future<bool> validateEncrypt() async {
     Uint8List encrypted = await aes.encrypt(hex.decode(plain));
     Uint8List decrypted = await aes.decrypt(encrypted);
-    return hex.encode(encrypted) == expectedEncrypted && hex.encode(decrypted) == plain;
-  }
-
-  Future<bool> validateIsolatedEncrypt() async {
-    Uint8List encrypted = await aes.encryptIsolated(hex.decode(plain));
-    Uint8List decrypted = await aes.decryptIsolated(encrypted);
-    return hex.encode(encrypted) == expectedEncrypted && hex.encode(decrypted) == plain;
+    return hex.encode(encrypted) == expectedEncrypted &&
+        hex.encode(decrypted) == plain;
   }
 }
 
-typedef HashMethod = Future<Uint8List> Function(Uint8List password, Uint8List salt);
+typedef HashMethod = Future<Uint8List> Function(
+    Uint8List password, Uint8List salt);
 
 abstract class HashTestCase extends TestCase {
   final String password;
   final String salt;
   final String expectedHashCode;
 
-  HashTestCase(String name, this.password, this.salt, this.expectedHashCode) : super(name);
+  HashTestCase(String name, this.password, this.salt, this.expectedHashCode)
+      : super(name);
 
   @override
   Future<bool> execute() async {
-    Uint8List hashed = await getHashMethod().call(hex.decode(password), hex.decode(salt));
-    Uint8List hashed1 = await getIsolatedHashMethod().call(hex.decode(password), hex.decode(salt));
-    return hex.encode(hashed) == hex.encode(hashed1) && hex.encode(hashed) == expectedHashCode;
+    Uint8List hashed =
+        await getHashMethod().call(hex.decode(password), hex.decode(salt));
+    return hex.encode(hashed) == expectedHashCode;
   }
 
   HashMethod getHashMethod();
-
-  HashMethod getIsolatedHashMethod();
 }
 
 class Argon2iTestCase extends HashTestCase {
   final Argon2 argon2;
 
-  Argon2iTestCase(String name, this.argon2, String password, String salt, String expectedHashCode)
+  Argon2iTestCase(String name, this.argon2, String password, String salt,
+      String expectedHashCode)
       : super(name, password, salt, expectedHashCode);
 
   @override
   HashMethod getHashMethod() => argon2.argon2i;
-
-  @override
-  HashMethod getIsolatedHashMethod() => argon2.argon2iIsolated;
 }
 
 class Argon2dTestCase extends HashTestCase {
   final Argon2 argon2;
 
-  Argon2dTestCase(String name, this.argon2, String password, String salt, String expectedHashCode)
+  Argon2dTestCase(String name, this.argon2, String password, String salt,
+      String expectedHashCode)
       : super(name, password, salt, expectedHashCode);
 
   @override
   HashMethod getHashMethod() => argon2.argon2d;
-
-  @override
-  HashMethod getIsolatedHashMethod() => argon2.argon2dIsolated;
 }
 
 class Argon2idTestCase extends HashTestCase {
   final Argon2 argon2;
 
-  Argon2idTestCase(String name, this.argon2, String password, String salt, String expectedHashCode)
+  Argon2idTestCase(String name, this.argon2, String password, String salt,
+      String expectedHashCode)
       : super(name, password, salt, expectedHashCode);
 
   @override
   HashMethod getHashMethod() => argon2.argon2id;
-
-  @override
-  HashMethod getIsolatedHashMethod() => argon2.argon2idIsolated;
 }
 
 class Report {