+++ CHANGELOG.md
@@ -12,12 +12,10 @@ Notice: version 0.0.1 is deprecated, consider using the latest version.
 * Now it's able to configure custom argon2 parameters
 * Added a test report UI in example, for which you can validate the result in real devices
 
-## 0.0.3 (Deprecated)
+## 0.0.3
 
 * Implicitly declare api return value type
 * Using separate thread to run android native code, thus it might help to un-block the ui thread
-
-Notice: the source code of this version is not available... Somehow I did not commit that.
 
 ## 0.0.2+1
 
@@ -37,3 +35,8 @@ Notice: the source code of this version is not available... Somehow I did not co
 
 * **NEW** `*Isloated` APIs are now available for running in separated isolate, since argon2 and AES may take long time to finish, which will then block the UI. Old APIs are still available.
 * Refactor examples
+
+## 0.0.6
+
+* Removed the `*Isloated` APIs introduced in 0.0.5, because the performance is not ideal.
+* Using native thread(in Android) and DispatchQueue(in ios) to unblock main UI